Deque 双端队列 对可叠加优惠信息进行排序

+++ gulimall-cart/src/main/java/com/atguigu/gulimall/cart/vo/SkuFullReductionVo.java
@@ -27,7 +27,7 @@ public class SkuFullReductionVo {
      * 满 fullCount 件，打 discount 折
      */
     private Integer fullCount;
-    private Integer discount;
+    private BigDecimal discount;
 
 
     /**
