+++ gulimall-sms/src/main/java/com/atguigu/gulimall/sms/to/SkuReductionTo.java
@@ -24,7 +24,7 @@ public class SkuReductionTo {
      * 满 fullCount 件，打 discount 折
      */
     private Integer fullCount;
-    private Integer discount;
+    private BigDecimal discount;
 
 
     /**
